# docs/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
paths:
...
  /payments:
    get:
      tags:
      - payment-controller
      operationId: listPayments
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentDto"
    post:
      tags:
      - payment-controller
      operationId: createPayment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentDto"
  /memberships:
    get:
      tags:
      - membership-controller
      operationId: listMemberships
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MembershipDto"
    post:
      tags:
      - membership-controller
      operationId: createMembership
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MembershipDto"
  /membership-types:
    get:
      tags:
      - membership-type-controller
      operationId: listMembershipTypes
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MembershipTypeDto"
    post:
      tags:
      - membership-type-controller
      operationId: createMembershipType
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipTypeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MembershipTypeDto"
...
  /halls:
    get:
      tags:
      - hall-controller
      operationId: listHalls
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HallDto"
    post:
      tags:
      - hall-controller
      operationId: createHall
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HallRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HallDto"
...
  /chats/{id}/participants:
    post:
      tags:
      - chat-controller
      operationId: addChatParticipant
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatParticipantCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChatParticipantDto"
...
  /workout-sessions/{id}:
    get:
      tags:
      - workout-session-controller
      operationId: getWorkoutSession
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkoutSessionDto"
    patch:
      tags:
      - workout-session-controller
      operationId: updateCard
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutSessionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkoutSessionDto"
...
  /payments/{id}:
    get:
      tags:
      - payment-controller
      operationId: getPayment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentDto"
    patch:
      tags:
      - payment-controller
      operationId: updatePayment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentDto"
  /memberships/{id}:
    get:
      tags:
      - membership-controller
      operationId: getMembership
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MembershipDto"
    patch:
      tags:
      - membership-controller
      operationId: patchMembership
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MembershipDto"
  /membership-types/{id}:
    get:
      tags:
      - membership-type-controller
      operationId: getMembershipType
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MembershipTypeDto"
    patch:
      tags:
      - membership-type-controller
      operationId: updateCard_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipTypeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MembershipTypeDto"
...
  /halls/{id}:
    get:
      tags:
      - hall-controller
      operationId: getHall
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HallDto"
    patch:
      tags:
      - hall-controller
      operationId: updateHall
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HallRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HallDto"
  /chats/{id}/participants/{userId}:
    patch:
      tags:
      - chat-controller
      operationId: updateChatParticipant
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatParticipantPatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChatParticipantDto"
  /cards/{id}:
    get:
      tags:
      - card-controller
      operationId: getCard
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
    patch:
      tags:
      - card-controller
      operationId: updateCard_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
...
  /chats/{id}:
    get:
      tags:
      - chat-controller
      operationId: getChat
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChatDto"
...
components:
  schemas:
    WorkoutSessionRequest:
      type: object
      properties:
        coachUuid:
          type: string
          format: uuid
        hallUuid:
          type: string
          format: uuid
        date:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: string
      required:
      - coachUuid
      - date
      - description
      - hallUuid
      - title
    Card:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        country:
          $ref: "#/components/schemas/Country"
        cardNumber:
          type: string
        cvv:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        nameOnCard:
          type: string
        postalCode:
          type: string
      required:
      - cardNumber
      - cvv
      - dateOfBirth
      - nameOnCard
      - postalCode
    Country:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        countryName:
          type: string
      required:
      - countryName
    Exercise:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
        targetMuscles:
          type: array
          items:
            $ref: "#/components/schemas/TargetMuscle"
          uniqueItems: true
    Hall:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        hallType:
          $ref: "#/components/schemas/HallType"
        hallName:
          type: string
        hallDescription:
          type: string
    HallType:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
    Membership:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        purchaseDate:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        membershipType:
          $ref: "#/components/schemas/MembershipType"
    MembershipType:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        type:
          type: string
        price:
          type: number
          format: float
    TargetMuscle:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
    User:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/UserRole"
        card:
          $ref: "#/components/schemas/Card"
        membership:
          $ref: "#/components/schemas/Membership"
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        email:
          type: string
        phoneNumber:
          type: string
        imageUrl:
          type: string
      required:
      - dateOfBirth
      - firstName
      - lastName
    UserRole:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        role:
          type: string
    WorkoutSessionDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        coach:
          $ref: "#/components/schemas/User"
        hall:
          $ref: "#/components/schemas/Hall"
        date:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: string
        attendants:
          type: array
          items:
            $ref: "#/components/schemas/User"
          uniqueItems: true
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSessionExerciseDto"
    WorkoutSessionExerciseDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        exercise:
          $ref: "#/components/schemas/Exercise"
        reps:
          type: integer
          format: int32
        exerciseOrder:
          type: integer
          format: int32
...
    PaymentRequest:
      type: object
      properties:
        cardUuid:
          type: string
          format: uuid
        statusUuid:
          type: string
          format: uuid
        membershipUuid:
          type: string
          format: uuid
      required:
      - cardUuid
      - membershipUuid
      - statusUuid
    PaymentDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        card:
          $ref: "#/components/schemas/Card"
        status:
          $ref: "#/components/schemas/PaymentStatus"
        membership:
          $ref: "#/components/schemas/Membership"
    PaymentStatus:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        status:
          type: string
    MembershipRequest:
      type: object
      properties:
        purchaseDate:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        membershipTypeUuid:
          type: string
          format: uuid
      required:
      - membershipTypeUuid
      - purchaseDate
      - validUntil
    MembershipDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        purchaseDate:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        membershipType:
          $ref: "#/components/schemas/MembershipType"
    MembershipTypeRequest:
      type: object
      properties:
        type:
          type: string
        price:
          type: number
          format: float
      required:
      - price
      - type
    MembershipTypeDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        type:
          type: string
        price:
          type: number
          format: float
...
    HallRequest:
      type: object
      properties:
        hallTypeUuid:
          type: string
          format: uuid
        hallName:
          type: string
        hallDescription:
          type: string
      required:
      - hallDescription
      - hallName
      - hallTypeUuid
    HallDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        hallType:
          $ref: "#/components/schemas/HallType"
        hallName:
          type: string
        hallDescription:
          type: string
...
    ChatDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ChatParticipantDto"
          uniqueItems: true
    ChatParticipantDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        user:
          $ref: "#/components/schemas/User"
        lastReadDateTime:
          type: string
          format: date-time
    ChatParticipantCreateRequest:
      type: object
      properties:
        userUuid:
          type: string
          format: uuid
        lastReadDateTime:
          type: string
          format: date-time
      required:
      - userUuid
    CardRequest:
      type: object
      properties:
        countryUuid:
          type: string
          format: uuid
        cardNumber:
          type: string
        cvv:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        nameOnCard:
          type: string
        postalCode:
          type: string
      required:
      - cardNumber
      - countryUuid
      - cvv
      - dateOfBirth
      - nameOnCard
      - postalCode
    CardDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        country:
          $ref: "#/components/schemas/Country"
        cardNumber:
          type: string
        cvv:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        nameOnCard:
          type: string
        postalCode:
          type: string
    ChatParticipantPatchRequest:
      type: object
      properties:
        lastReadDateTime:
          type: string
          format: date-time
      required:
      - lastReadDateTime
...